+++ gtfs_kit/trips.py
@@ -38,10 +38,10 @@ def is_active_trip(feed: "Feed", trip_id: str, date: str) -> bool:
     """
     service = feed._trips_i.at[trip_id, "service_id"]
     # Check feed._calendar_dates_g.
-    caldg = feed._calendar_dates_g
-    if caldg is not None:
-        if (service, date) in caldg.groups:
-            et = caldg.get_group((service, date))["exception_type"].iat[0]
+    caldi = feed._calendar_dates_i
+    if caldi is not None:
+        if (service, date) in caldi.index:
+            et = caldi.at[(service, date), "exception_type"]
             if et == 1:
                 return True
             else: